ADD: fix norm and add add errors into lexer level
